creative-forge/creative-forge/requirements.txt
fastapi
python-dotenv
langchain
langchain-openai
langchain-mcp-adapters
fastmcp
swagger-ui-bundle
uvicorn
httpx
